Add TungstenFabric compatibility to charts
The charts changes are required for deployment of various clouds based on Tungsten Fabric SDN. Right now it's tested for Airship-in-a-bottle.

The code cannot be tested currently in OpenStack Helm project because of absence of tests and platform for that.

This patchset doesn't have Heat-related changes, they'll be added later.
Change-Id: I73f2ced2b09dbb93146334b59fe4571fa13dbfb0